lambda/lambda-auth/requirements.txt
google-api-python-client==2.137.0
